+++ philo/sources/exit.c
@@ -6,7 +6,7 @@
 /*   By: yukravch <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/01 12:03:06 by yukravch          #+#    #+#             */
-/*   Updated: 2025/05/06 14:57:44 by yukravch         ###   ########.fr       */
+/*   Updated: 2025/05/06 15:20:20 by yukravch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@ void	ft_error(char *msg)
 	write(2, "\n", 1);
 }
 
-void    ft_free_array(t_philos** philos, int nb_of_philos)
+void    ft_free_array(t_philos** philos, size_t nb_of_philos)
 {
 
-        int     i = 0;
+        size_t    i = 0;
         while (i < nb_of_philos && philos[i])
         {
                 free(philos[i]);
